Working chat with firebase
