+++ client/src/index.jsx
@@ -12,7 +12,7 @@ class App extends React.Component {
   constructor(props) {
     super(props)
     this.state = {
-      isLogin: false,
+      isLogin: true,
       userData: userProfileDummyData
     }
     this.toggleLogin = this.toggleLogin.bind(this)
